backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# AI Services
openai>=1.3.0
google-generativeai>=0.3.0

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0

# Document Processing
PyPDF2>=3.0.0
python-magic-bin>=0.4.14; sys_platform == 'win32'
pillow>=10.1.0

# NLP
spacy>=3.7.0

# Utilities
tqdm>=4.66.0
